refactor: Rename, reorganize schema module (#1963)
