+++ ludwig/encoders/sequence_encoders.py
@@ -21,7 +21,7 @@ from torch import nn
 
 from ludwig.constants import AUDIO, SEQUENCE, TEXT, TIMESERIES
 from ludwig.encoders.base import Encoder
-from ludwig.encoders.registry import register_encoder
+from ludwig.encoders.registry import register_encoder, register_sequence_encoder
 from ludwig.modules.attention_modules import TransformerStack
 from ludwig.modules.convolutional_modules import Conv1DStack, ParallelConv1D, ParallelConv1DStack
 from ludwig.modules.embedding_modules import EmbedSequence, TokenAndPositionEmbedding
@@ -204,6 +204,7 @@ class SequenceEmbedEncoder(Encoder):
         return self.reduce_sequence.output_shape
 
 
+@register_sequence_encoder("parallel_cnn")
 @register_encoder("parallel_cnn", [AUDIO, SEQUENCE, TEXT, TIMESERIES])
 class ParallelCNN(Encoder):
     def __init__(
@@ -496,6 +497,7 @@ class ParallelCNN(Encoder):
         return self.parallel_conv1d.output_shape
 
 
+@register_sequence_encoder("stacked_cnn")
 @register_encoder("stacked_cnn", [AUDIO, SEQUENCE, TEXT, TIMESERIES])
 class StackedCNN(Encoder):
     def __init__(
@@ -825,6 +827,7 @@ class StackedCNN(Encoder):
         return {"encoder_output": hidden}
 
 
+@register_sequence_encoder("stacked_parallel_cnn")
 @register_encoder("stacked_parallel_cnn", [AUDIO, SEQUENCE, TEXT, TIMESERIES])
 class StackedParallelCNN(Encoder):
     def __init__(
@@ -1129,6 +1132,7 @@ class StackedParallelCNN(Encoder):
         return {"encoder_output": hidden}
 
 
+@register_sequence_encoder("rnn")
 @register_encoder("rnn", [AUDIO, SEQUENCE, TEXT, TIMESERIES])
 class StackedRNN(Encoder):
     def __init__(
@@ -1394,6 +1398,7 @@ class StackedRNN(Encoder):
         return {"encoder_output": hidden, "encoder_output_state": final_state}
 
 
+@register_sequence_encoder("cnnrnn")
 @register_encoder("cnnrnn", [AUDIO, SEQUENCE, TEXT, TIMESERIES])
 class StackedCNNRNN(Encoder):
     def __init__(
