+++ ludwig/schema/combiners/__init__.py
@@ -1,0 +1,9 @@
+# Pull configs here for convenient imports in other modules:
+from ludwig.schema.combiners.comparator import ComparatorCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.concat import ConcatCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.project_aggregate import ProjectAggregateCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.sequence import SequenceCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.sequence_concat import SequenceConcatCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.tab_transformer import TabTransformerCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.tabnet import TabNetCombinerConfig  # noqa: F401
+from ludwig.schema.combiners.transformer import TransformerCombinerConfig  # noqa: F401
