+++ tests/integration_tests/test_custom_components.py
@@ -15,6 +15,7 @@ from ludwig.encoders.base import Encoder
 from ludwig.encoders.registry import register_encoder
 from ludwig.modules.loss_modules import LogitsInputsMixin, register_loss
 from ludwig.modules.metric_modules import LossMetric, register_metric
+from ludwig.schema.combiners.base import BaseCombinerConfig
 from tests.integration_tests.utils import (
     category_feature,
     generate_data,
