+++ tests/integration_tests/utils.py
@@ -660,7 +660,9 @@ def train_with_backend(
                     assert k1 == k2
                     for (name1, metric1), (name2, metric2) in zip(v1.items(), v2.items()):
                         assert name1 == name2
-                        assert np.isclose(metric1, metric2, rtol=1e-04), f"metric {name1}: {metric1} != {metric2}"
+                        assert np.isclose(
+                            metric1, metric2, rtol=1e-04, atol=1e-5
+                        ), f"metric {name1}: {metric1} != {metric2}"
 
         return model
     finally:
