+++ tests/ludwig/marshmallow/test_fields_misc.py
@@ -4,7 +4,7 @@ import pytest
 from marshmallow.exceptions import ValidationError as MarshmallowValidationError
 from marshmallow_dataclass import dataclass
 
-import ludwig.marshmallow.marshmallow_schema_utils as lusutils
+from ludwig.schema import utils as schema_utils
 
 
 def get_marshmallow_from_dataclass_field(dfield):
@@ -19,14 +19,14 @@ def test_StringOptions():
     # Test case of default conflicting with allowed options:
     test_options = ["one"]
     with pytest.raises(MarshmallowValidationError):
-        lusutils.StringOptions(test_options, default=None, nullable=False)
+        schema_utils.StringOptions(test_options, default=None, nullable=False)
 
     # Test creating a schema with simple option, null not allowed:
     test_options = ["one"]
 
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: str = lusutils.StringOptions(test_options, "one", nullable=False)
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: str = schema_utils.StringOptions(test_options, "one", nullable=False)
 
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": None})
@@ -37,14 +37,14 @@ def test_StringOptions():
 
 def test_Embed():
     # Test metadata matches expected defaults after field creation (null allowed):
-    default_embed = get_marshmallow_from_dataclass_field(lusutils.Embed())
+    default_embed = get_marshmallow_from_dataclass_field(schema_utils.Embed())
     assert default_embed.default is None
     assert default_embed.allow_none is True
 
     # Test simple schema creation:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Union[None, str, int] = lusutils.Embed()
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Union[None, str, int] = schema_utils.Embed()
 
     # Test null/empty loading cases:
     assert CustomTestSchema.Schema().load({}).foo is None
@@ -63,20 +63,20 @@ def test_Embed():
 
 def test_InitializerOrDict():
     # Test metadata matches expected defaults after field creation (null allowed):
-    default_initializerordict = get_marshmallow_from_dataclass_field(lusutils.InitializerOrDict())
+    default_initializerordict = get_marshmallow_from_dataclass_field(schema_utils.InitializerOrDict())
     assert default_initializerordict.default == "xavier_uniform"
 
-    initializerordict = get_marshmallow_from_dataclass_field(lusutils.InitializerOrDict("zeros"))
+    initializerordict = get_marshmallow_from_dataclass_field(schema_utils.InitializerOrDict("zeros"))
     assert initializerordict.default == "zeros"
 
     # Test default value validation:
     with pytest.raises(MarshmallowValidationError):
-        lusutils.InitializerOrDict("test")
+        schema_utils.InitializerOrDict("test")
 
     # Test simple schema creation:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Union[None, str, Dict] = lusutils.InitializerOrDict()
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Union[None, str, Dict] = schema_utils.InitializerOrDict()
 
     # Test invalid non-dict loads:
     with pytest.raises(MarshmallowValidationError):
@@ -103,17 +103,17 @@ def test_InitializerOrDict():
 
 def test_FloatRangeTupleDataclassField():
     # Test metadata matches expected defaults after field creation (null not allowed):
-    default_floatrange_tuple = get_marshmallow_from_dataclass_field(lusutils.FloatRangeTupleDataclassField())
+    default_floatrange_tuple = get_marshmallow_from_dataclass_field(schema_utils.FloatRangeTupleDataclassField())
     assert default_floatrange_tuple.default == (0.9, 0.999)
 
     # Test dimensional mismatch:
     with pytest.raises(MarshmallowValidationError):
-        lusutils.FloatRangeTupleDataclassField(N=3, default=(1, 1))
+        schema_utils.FloatRangeTupleDataclassField(N=3, default=(1, 1))
 
     # Test default schema creation:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Tuple[float, float] = lusutils.FloatRangeTupleDataclassField()
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Tuple[float, float] = schema_utils.FloatRangeTupleDataclassField()
 
     # Test empty load:
     assert CustomTestSchema.Schema().load({}).foo == (0.9, 0.999)
@@ -128,8 +128,8 @@ def test_FloatRangeTupleDataclassField():
 
     # Test non-default schema (N=3, other custom metadata):
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Tuple[float, float] = lusutils.FloatRangeTupleDataclassField(N=3, default=(1, 1, 1), min=-10, max=10)
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Tuple[float, float] = schema_utils.FloatRangeTupleDataclassField(N=3, default=(1, 1, 1), min=-10, max=10)
 
     assert CustomTestSchema.Schema().load({}).foo == (1, 1, 1)
     assert CustomTestSchema.Schema().load({"foo": [2, 2, 2]}).foo == (2, 2, 2)
