+++ tests/ludwig/marshmallow/test_fields_optimization.py
@@ -5,8 +5,8 @@ import pytest
 from marshmallow.exceptions import ValidationError as MarshmallowValidationError
 from marshmallow_dataclass import dataclass
 
-import ludwig.marshmallow.marshmallow_schema_utils as lusutils
-import ludwig.modules.optimization_modules as lmo
+import ludwig.schema.optimizers as lso
+from ludwig.schema import utils as schema_utils
 
 # Tests for custom dataclass/marshmallow fields:
 
@@ -17,9 +17,9 @@ def get_marshmallow_from_dataclass_field(dfield):
 
 
 def test_torch_description_pull():
-    example_empty_desc_prop = lusutils.unload_jsonschema_from_marshmallow_class(lmo.AdamOptimizerConfig)["properties"][
-        "lr"
-    ]
+    example_empty_desc_prop = schema_utils.unload_jsonschema_from_marshmallow_class(lso.AdamOptimizerConfig)[
+        "properties"
+    ]["lr"]
     assert (
         isinstance(example_empty_desc_prop, dict)
         and "description" in example_empty_desc_prop
@@ -30,46 +30,46 @@ def test_torch_description_pull():
 
 def test_OptimizerDataclassField():
     # Test default case:
-    default_optimizer_field = lmo.OptimizerDataclassField()
+    default_optimizer_field = lso.OptimizerDataclassField()
     assert default_optimizer_field.default_factory is not None
     assert get_marshmallow_from_dataclass_field(default_optimizer_field).allow_none is False
-    assert default_optimizer_field.default_factory() == lmo.AdamOptimizerConfig()
+    assert default_optimizer_field.default_factory() == lso.AdamOptimizerConfig()
 
     # Test normal cases:
-    optimizer_field = lmo.OptimizerDataclassField({"type": "adamax"})
+    optimizer_field = lso.OptimizerDataclassField({"type": "adamax"})
     assert optimizer_field.default_factory is not None
     assert get_marshmallow_from_dataclass_field(optimizer_field).allow_none is False
-    assert optimizer_field.default_factory() == lmo.AdamaxOptimizerConfig()
+    assert optimizer_field.default_factory() == lso.AdamaxOptimizerConfig()
 
-    optimizer_field = lmo.OptimizerDataclassField({"type": "adamax", "betas": (0.1, 0.1)})
+    optimizer_field = lso.OptimizerDataclassField({"type": "adamax", "betas": (0.1, 0.1)})
     assert optimizer_field.default_factory is not None
     assert get_marshmallow_from_dataclass_field(optimizer_field).allow_none is False
     assert optimizer_field.default_factory().betas == (0.1, 0.1)
 
     # Test invalid default case:
     with pytest.raises(MarshmallowValidationError):
-        lmo.OptimizerDataclassField({})
+        lso.OptimizerDataclassField({})
     with pytest.raises(MarshmallowValidationError):
-        lmo.OptimizerDataclassField("test")
+        lso.OptimizerDataclassField("test")
     with pytest.raises(MarshmallowValidationError):
-        lmo.OptimizerDataclassField(None)
+        lso.OptimizerDataclassField(None)
     with pytest.raises(MarshmallowValidationError):
-        lmo.OptimizerDataclassField(1)
+        lso.OptimizerDataclassField(1)
 
     # Test creating a schema with default options:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Optional[lmo.BaseOptimizerConfig] = lmo.OptimizerDataclassField()
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Optional[lso.BaseOptimizerConfig] = lso.OptimizerDataclassField()
 
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": "test"})
 
-    assert CustomTestSchema.Schema().load({}).foo == lmo.AdamOptimizerConfig()
+    assert CustomTestSchema.Schema().load({}).foo == lso.AdamOptimizerConfig()
 
     # Test creating a schema with set default:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Optional[lmo.BaseOptimizerConfig] = lmo.OptimizerDataclassField({"type": "adamax", "betas": (0.1, 0.1)})
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Optional[lso.BaseOptimizerConfig] = lso.OptimizerDataclassField({"type": "adamax", "betas": (0.1, 0.1)})
 
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": None})
@@ -78,65 +78,65 @@ def test_OptimizerDataclassField():
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": {"type": "invalid", "betas": (0.2, 0.2)}})
 
-    assert CustomTestSchema.Schema().load({}).foo == lmo.AdamaxOptimizerConfig(betas=(0.1, 0.1))
+    assert CustomTestSchema.Schema().load({}).foo == lso.AdamaxOptimizerConfig(betas=(0.1, 0.1))
     assert CustomTestSchema.Schema().load(
         {"foo": {"type": "adamax", "betas": (0.2, 0.2)}}
-    ).foo == lmo.AdamaxOptimizerConfig(betas=(0.2, 0.2))
+    ).foo == lso.AdamaxOptimizerConfig(betas=(0.2, 0.2))
     assert CustomTestSchema.Schema().load(
         {"foo": {"type": "adamax", "betas": (0.2, 0.2), "extra_key": 1}}
-    ).foo == lmo.AdamaxOptimizerConfig(betas=(0.2, 0.2))
+    ).foo == lso.AdamaxOptimizerConfig(betas=(0.2, 0.2))
 
 
 def test_ClipperDataclassField():
     # Test default case:
-    default_clipper_field = lmo.GradientClippingDataclassField()
+    default_clipper_field = lso.GradientClippingDataclassField()
     assert default_clipper_field.default_factory is not None
     assert get_marshmallow_from_dataclass_field(default_clipper_field).allow_none is True
-    assert default_clipper_field.default_factory() == lmo.GradientClippingConfig()
+    assert default_clipper_field.default_factory() == lso.GradientClippingConfig()
 
     # Test normal cases:
-    clipper_field = lmo.GradientClippingDataclassField({"clipglobalnorm": 0.1})
+    clipper_field = lso.GradientClippingDataclassField({"clipglobalnorm": 0.1})
     assert clipper_field.default_factory is not None
     assert get_marshmallow_from_dataclass_field(clipper_field).allow_none is True
-    assert clipper_field.default_factory() == lmo.GradientClippingConfig(clipglobalnorm=0.1)
+    assert clipper_field.default_factory() == lso.GradientClippingConfig(clipglobalnorm=0.1)
 
-    clipper_field = lmo.GradientClippingDataclassField({"clipglobalnorm": None})
+    clipper_field = lso.GradientClippingDataclassField({"clipglobalnorm": None})
     assert clipper_field.default_factory is not None
     assert get_marshmallow_from_dataclass_field(clipper_field).allow_none is True
-    assert clipper_field.default_factory() == lmo.GradientClippingConfig(clipglobalnorm=None)
+    assert clipper_field.default_factory() == lso.GradientClippingConfig(clipglobalnorm=None)
 
     # Test invalid default case:
     with pytest.raises(MarshmallowValidationError):
-        lmo.GradientClippingDataclassField("test")
+        lso.GradientClippingDataclassField("test")
     with pytest.raises(MarshmallowValidationError):
-        lmo.GradientClippingDataclassField(None)
+        lso.GradientClippingDataclassField(None)
     with pytest.raises(MarshmallowValidationError):
-        lmo.GradientClippingDataclassField(1)
+        lso.GradientClippingDataclassField(1)
 
     # Test creating a schema with default options:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Optional[lmo.GradientClippingConfig] = lmo.GradientClippingConfig()
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Optional[lso.GradientClippingConfig] = lso.GradientClippingConfig()
 
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": "test"})
 
-    assert CustomTestSchema.Schema().load({}).foo == lmo.GradientClippingConfig()
+    assert CustomTestSchema.Schema().load({}).foo == lso.GradientClippingConfig()
 
     # Test creating a schema with set default:
     @dataclass
-    class CustomTestSchema(lusutils.BaseMarshmallowConfig):
-        foo: Optional[lmo.GradientClippingConfig] = lmo.GradientClippingDataclassField({"clipglobalnorm": 0.1})
+    class CustomTestSchema(schema_utils.BaseMarshmallowConfig):
+        foo: Optional[lso.GradientClippingConfig] = lso.GradientClippingDataclassField({"clipglobalnorm": 0.1})
 
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": "test"})
     with pytest.raises(MarshmallowValidationError):
         CustomTestSchema.Schema().load({"foo": {"clipglobalnorm": "invalid"}})
 
-    assert CustomTestSchema.Schema().load({}).foo == lmo.GradientClippingConfig(clipglobalnorm=0.1)
-    assert CustomTestSchema.Schema().load({"foo": {"clipglobalnorm": 1}}).foo == lmo.GradientClippingConfig(
+    assert CustomTestSchema.Schema().load({}).foo == lso.GradientClippingConfig(clipglobalnorm=0.1)
+    assert CustomTestSchema.Schema().load({"foo": {"clipglobalnorm": 1}}).foo == lso.GradientClippingConfig(
         clipglobalnorm=1
     )
     assert CustomTestSchema.Schema().load(
         {"foo": {"clipglobalnorm": 1, "extra_key": 1}}
-    ).foo == lmo.GradientClippingConfig(clipglobalnorm=1)
+    ).foo == lso.GradientClippingConfig(clipglobalnorm=1)
