+++ tests/ludwig/marshmallow/test_marshmallow_misc.py
@@ -4,9 +4,8 @@ from marshmallow.utils import EXCLUDE
 from marshmallow_dataclass import dataclass
 
 import ludwig.combiners.combiners as lcc
-import ludwig.marshmallow.marshmallow_schema_utils as lusutils
-from ludwig.marshmallow.marshmallow_schema_utils import BaseMarshmallowConfig
 from ludwig.models.trainer import TrainerConfig
+from ludwig.schema.utils import assert_is_a_marshmallow_class, BaseMarshmallowConfig, load_config_with_kwargs
 
 
 @dataclass
@@ -18,9 +17,9 @@ class CustomTestSchema(BaseMarshmallowConfig):
 
 
 def test_assert_is_a_marshmallow_clas():
-    lusutils.assert_is_a_marshmallow_class(TrainerConfig)
+    assert_is_a_marshmallow_class(TrainerConfig)
     with pytest.raises(AssertionError, match=r"^Expected marshmallow class.*"):
-        lusutils.assert_is_a_marshmallow_class(lcc.ConcatCombiner)
+        assert_is_a_marshmallow_class(lcc.ConcatCombiner)
 
 
 def test_custom_marshmallow_inheritance():
@@ -32,18 +31,18 @@ def test_load_config_with_kwargs():
         "foo": 6,
         "bar": 6,
     }
-    initialized_class, leftover = lusutils.load_config_with_kwargs(CustomTestSchema, test_kwargs)
+    initialized_class, leftover = load_config_with_kwargs(CustomTestSchema, test_kwargs)
 
     assert initialized_class.foo == 6
     assert leftover == {"bar": 6}
 
     # TransformerCombiner has no required/non-default arguments:
-    initialized_class, leftover = lusutils.load_config_with_kwargs(lcc.TransformerCombinerConfig, test_kwargs)
+    initialized_class, leftover = load_config_with_kwargs(lcc.TransformerCombinerConfig, test_kwargs)
     assert initialized_class.bias_initializer == "zeros"
     assert leftover == test_kwargs
-    initialized_class, leftover = lusutils.load_config_with_kwargs(lcc.TransformerCombinerConfig, {})
+    initialized_class, leftover = load_config_with_kwargs(lcc.TransformerCombinerConfig, {})
     assert leftover == {}
 
     # ComparatorCombiner does have required arguments, so expect a failure:
     with pytest.raises(MarshmallowValidationError):
-        initialized_class, leftover = lusutils.load_config_with_kwargs(lcc.ComparatorCombinerConfig, test_kwargs)
+        initialized_class, leftover = load_config_with_kwargs(lcc.ComparatorCombinerConfig, test_kwargs)
